Refactoring and architecture
