--- fileserver.h
@@ -1,45 +1,0 @@
-#ifndef FILESERVER_H
-#define FILESERVER_H
-
-#include <QObject>
-#include <QTcpServer>
-#include <QTcpSocket>
-
-#include <QFile>
-#include <QDir>
-#include <QByteArray>
-
-#include <QJsonObject>
-#include <QJsonDocument>
-
-#include <QDateTime>
-
-#include <QDebug>
-#include <QUuid>
-
-class FileServer : public QTcpServer
-{
-    Q_OBJECT
-public:
-    explicit FileServer(QObject *parent = nullptr);
-
-signals:
-    void saveFileToDatabase(const QString& fileUrl);
-    void setAvatarInDatabase(const QString& avatarUrl, const int& user_id);
-    void sendVoiceMessage(QJsonObject voiceJson);
-protected:
-    void incomingConnection(qintptr handle) Q_DECL_OVERRIDE;
-private slots:
-    void sendData(QJsonDocument& sendDoc);
-    void readClient();
-    void makeAvatarUrlProcessing(const QJsonObject& json);
-    QString makeUrlProcessing(const QJsonObject& json);
-    void getFileFromUrlProcessing(const QString &fileUrl, const QString &flag);
-    void getAvatarFromServer(const QJsonObject &json);
-    void voiceMessageProcessing(QJsonObject &voiceJson);
-
-private:
-    QTcpSocket* socket;
-};
-
-#endif // FILESERVER_H
